requirements.txt
invoke
pytest
pytest-cov
pytest-pep8
flask
redis
mock
mockredispy